Put all the cache tests back into one file
